get/update rates routes + config
------------------------------------------------------------------------
GetRatesRoute - gets all existing rates

CreateRateRoute - creates a new rate if it does not conflict with any
existring rates

OverwriteRatesRoute - overwrites all existing rates

------------------------------------------------------------------------
These routes are supported by new input/output types defined in
pkg\types

------------------------------------------------------------------------
To create a rate, we validate for a non-zero price, loadable timezone,
a comma separated list of days that have no repeated items, a timespan
specified as "HHMM-HHMM", and that there are no currently existing rates
for which the new rate would overlap.

To overwrite the rates, we check the same fields as creating a rate, but
leave off the overlap check since we are overwriting the rates. (If
you're wondering this is more of a toy function to make it easier to
load in a large amount of rates at once to test the functionality of the
app)

------------------------------------------------------------------------
the app now runs init() in internal\config\config.go which connects to
a dynamo db that contains the rates table