GetTimespanPriceRoute
------------------------------------------------------------------------
This route returns a price if one exists on a rate that covers a time
for which the given Start/End string falls within

We validate the Start/End string to be a true time range that parses
into time.RFC3339 (ISO-8601 equivalent), is within the same year, on the
same day, being longer than 0s, and having the same timezone offset

Then we try to match the timespan to an existing rate (line 116
internal\helpers\rates.go) - this was an especially interesting
function to write because of all the manipulation needed to get the
rate data in terms of the given time range. I have to say,
matchTimespanToRate (line 138 internal\helpers\util.go) was the best
part of this project.

This route is supported by input/output types defined in pkg\types.