Route Metrics
Route metrics include: AvgResponseTime, HitCount, SuccessCount,
and FailureCount. (see pkg\types\routemetrics.go)

Two helpers are now used in each route: UpdateRouteResponseTime and
UpdateRouteSuccessFailureCount. These are defered to the return of the
function and update the avg response time and the success or failure
count of the routes respectively. These run off of pre-defined route
names in internal\helpers\util.go, if new routes are written, new route
names should be defined there.

RouteMetricsTable and RouteMetricsTableConn are added to the config.

In some of the routes, I was assigning from the call to helpers directly
into the output. This was not ideal because it might lead to a client
consuming the output when they should really check for an error. I opt
to assign to the output only after all errors have returned nil.