Testing
Implement tests for functions in internal\helpers\validate and \util.
These functions contain most of the business logic and do not need a
DB connection nor an HTTP request to test.

Because the project used the special cmd folder to have the seeder
and server main functions, a blank main had to be created in the root to
allow for testing, otherwise Go throws a compile error.

Connection to DB tables was moved into the main func for seeder and
server because running tests runs the init() function. This caused an error
when no dynamo container was running at the 8000 port.

The Dockerfile was updated to run go test only in the directory that has
any tests (\internal\helpers\...)